+++ src/stylesheets/components/Header.css
@@ -1,20 +1,3 @@
-.navBar {
-  justify-content: space-between;
-  font-size: 1.3rem;
-  height: 5rem;
-  padding: 1rem;
-  box-shadow: 0 2px 12px rgba(0, 0, 0, 0.3);
-}
-
-.navBar a {
-  font-size: 1.2rem;
-}
-
-.navBar div {
-  align-items: center;
-  gap: 2rem;
-}
-
 .infoSection {
   margin: 2rem 1.3rem;
   gap: 2rem;
@@ -40,37 +23,4 @@
   height: 125px;
   object-fit: cover;
   clip-path: polygon(45% 1.33975%, 46.5798% 0.60307%, 48.26352% 0.15192%, 50% 0, 51.73648% 0.15192%, 53.4202% 0.60307%, 55% 1.33975%, 89.64102% 21.33975%, 91.06889% 22.33956%, 92.30146% 23.57212%, 93.30127% 25%, 94.03794% 26.5798%, 94.48909% 28.26352%, 94.64102% 30%, 94.64102% 70%, 94.48909% 71.73648%, 94.03794% 73.4202%, 93.30127% 75%, 92.30146% 76.42788%, 91.06889% 77.66044%, 89.64102% 78.66025%, 55% 98.66025%, 53.4202% 99.39693%, 51.73648% 99.84808%, 50% 100%, 48.26352% 99.84808%, 46.5798% 99.39693%, 45% 98.66025%, 10.35898% 78.66025%, 8.93111% 77.66044%, 7.69854% 76.42788%, 6.69873% 75%, 5.96206% 73.4202%, 5.51091% 71.73648%, 5.35898% 70%, 5.35898% 30%, 5.51091% 28.26352%, 5.96206% 26.5798%, 6.69873% 25%, 7.69854% 23.57212%, 8.93111% 22.33956%, 10.35898% 21.33975%);
-}
-
-.skills {
-  gap: 1rem;
-  width: 90%;
-}
-
-.level {
-  color: var(--secondary);
-}
-
-.icons {
-  font-size: 1.2rem;
-}
-
-.technologies {
-  padding: 1.4rem 0;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-}
-
-.technologies li button {
-  background: none;
-  border: none;
-  background-color: var(--gray-button);
-  padding: 0.5rem 1rem;
-  border-radius: 1rem;
-}
-
-.technologies li button:hover {
-  cursor: pointer;
-  background-color: var(--secondary);
-  color: var(--background);
 }
