+++ src/stylesheets/components/NavBar.css
@@ -1,0 +1,16 @@
+.navBar {
+  justify-content: space-between;
+  font-size: 1.3rem;
+  height: 5rem;
+  padding: 1rem;
+  box-shadow: 0 2px 12px rgba(0, 0, 0, 0.3);
+}
+
+.navBar a {
+  font-size: 1.2rem;
+}
+
+.navBar div {
+  align-items: center;
+  gap: 2rem;
+}
